requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.10.6
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.0

# Blockchain
web3==6.15.1
eth-account==0.9.0
requests==2.31.0